+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 * Replaced old user bulk processing with new implementation using the abstract bulk processing system.
 * `conditionContainers` template event in template `noticeAdd.tpl` added.
 * Use condition system for user search.
+* Image proxy for images included with the image BBCode.
 * Overhauled Redactor integration
 	* Linebreaks mode instead of using paragraphs, works better with the PHP-side parser which works with linebreaks
 	* Ported the PHP-BBCode parser, massively improves accuracy and ensures validity